requirements-dev.txt
# Dev Machine Requirements - ZeroMQ VLM Navigation + Neural Network Training
# Install in conda environment: pip install -r requirements-dev.txt

# Core dependencies
opencv-python==4.8.1.78
numpy==1.24.3

# VLM/AI dependencies  
google-generativeai==0.3.2
python-dotenv==1.0.0
Pillow==10.0.1

# PyTorch with CUDA (for neural network training)
--index-url https://download.pytorch.org/whl/cu124
torch==2.4.0+cu124
torchvision==0.19.0+cu124
torchaudio==2.4.0+cu124

# Neural Network Training dependencies
pandas>=1.5.0
seaborn>=0.12.0
scikit-learn>=1.3.0
tqdm>=4.65.0
tensorboard>=2.13.0

# Visualization and analysis
matplotlib==3.7.2

# ZeroMQ communication
pyzmq>=25.0.0

# Optional for joystick control
pygame>=2.1.0 

# Networking
requests==2.31.0